+++ src/commands/add.js
@@ -1,6 +1,6 @@
 const User = require("../../models/UserSchema");
 const moment = require("moment");
-const utils = require("../../utils/utils");
+const utils = require("../utils/utils");
 
 function createProperty(currentWins = 0, game) {
   // Use the matched games 'property' value